+++ src/KnownFileList.h
@@ -47,7 +47,7 @@ public:
 	bool	Init();
 	void	Save();
 	void	Clear();
-	CKnownFile*	FindKnownFile(wxString filename,uint32 in_date,uint32 in_size);
+	CKnownFile*	FindKnownFile(wxString filename,uint32 in_date,uint64 in_size);
 	CKnownFile*   FindKnownFileByID(const CMD4Hash& hash);
 	bool	IsKnownFile(const CKnownFile* file);
 
@@ -60,7 +60,7 @@ private:
 
 	bool	Append(CKnownFile*);
 
-	CKnownFile* IsOnDuplicates(wxString filename,uint32 in_date,uint32 in_size) const;
+	CKnownFile* IsOnDuplicates(wxString filename,uint32 in_date,uint64 in_size) const;
 
 	typedef std::list<CKnownFile*> KnownFileList;
 	KnownFileList	m_duplicates;
